// src/app/shared/quote.interface.ts
export interface Quote {
    cotizacion: string;
    producto: string;
    marca: string;
    referencia: string;
    cantidad: string;
    descripcion: string;
    fecha: string;
    precio: string;
    flete: string;
    total?: number;
    imagen: string;
    destinatario: string;
    solicitud: string;
    vendedor: string;
    sector: string;
    categoria: string;
    subcategoria: string;
    fileRef: string;
}